Fixed filename mismatch bug
We currently store streams for a video in a folder named after the video
title (which is itself stored under FILE_PATH). We sanitize video titles
to remove most special characters. It turns out that pytube does the
same thing with its default filename, so if you don't specify a filename
when downloading a video, it still has a valid filename for your OS.

There was a mismatch between our sanitized filename and their sanitized
filename, which led to a bug where the stream couldn't be found (the
wrong filename was being used). For example, brackets are valid
characters for folders/files, but we remove these; pytube does not with
its default filename. So we would be looking for '1234.mp4' when the
actual filename was '[1234].mp4'.

This has been resolved by specifying the
filename to be our sanitized filename, not the default.
I also removed filename_prefix as an argument since it's redundant
if we're going to be specifying the filename anyway.