Added template for config file
config_template.json contains a template for the config.json file used
by main.py to set configuration. If changes are made to the
configuration file (e.g. new options, removal/renaming of options), they
should be made to config_template.json. This way we can both have our
own personal config settings and not overwrite them when changes need to
be made to the layout of the config file.